+++ includes/parser.js
@@ -22,8 +22,12 @@ const pkg = require('../package.json');
 function parseBackupArgs() {
   var program = require('commander');
 
-  var defaults = config.cliDefaults();
-  config.applyEnvironmentVariables(defaults);
+  // Option CLI defaults
+  const defaults = config.cliDefaults();
+
+  // Options set by environment variables
+  const envVarOptions = {};
+  config.applyEnvironmentVariables(envVarOptions);
 
   program
     .version(pkg.version)
@@ -31,53 +35,61 @@ function parseBackupArgs() {
     .usage('[options...]')
     .option('-b, --buffer-size <n>',
       cliutils.getUsage('number of documents fetched at once', defaults.bufferSize),
-      Number, defaults.bufferSize)
+      Number)
     .option('-d, --db <db>',
-      cliutils.getUsage('name of the database to backup', defaults.db),
-      defaults.db)
+      cliutils.getUsage('name of the database to backup', defaults.db))
     .option('-k, --iam-api-key <API key>',
-      cliutils.getUsage('IAM API key to access the Cloudant server'),
-      defaults.iamApiKey)
+      cliutils.getUsage('IAM API key to access the Cloudant server'))
     .option('-l, --log <file>',
-      cliutils.getUsage('file to store logging information during backup', 'a temporary file'),
-      path.normalize, defaults.log)
+      cliutils.getUsage('file to store logging information during backup; invalid in "shallow" mode', 'a temporary file'),
+      path.normalize)
     .option('-m, --mode <mode>',
       cliutils.getUsage('"shallow" if only a superficial backup is done (ignoring conflicts and revision tokens), else "full" for complete backup', defaults.mode),
-      (mode) => { return mode.toLowerCase(); }, defaults.mode)
+      (mode) => { return mode.toLowerCase(); })
     .option('-o, --output <file>',
       cliutils.getUsage('file name to store the backup data', 'stdout'),
-      path.normalize, defaults.output)
+      path.normalize)
     .option('-p, --parallelism <n>',
-      cliutils.getUsage('number of HTTP requests to perform in parallel when performing a backup', defaults.parallelism),
-      Number, defaults.parallelism)
+      cliutils.getUsage('number of HTTP requests to perform in parallel when performing a backup; ignored in "shallow" mode', defaults.parallelism),
+      Number)
     .option('-r, --resume',
-      cliutils.getUsage('continue a previous backup from its last known position', defaults.resume),
-      defaults.resume)
+      cliutils.getUsage('continue a previous backup from its last known position; invalid in "shallow" mode', defaults.resume))
     .option('-t, --request-timeout <n>',
       cliutils.getUsage('milliseconds to wait for a response to a HTTP request before retrying the request', defaults.requestTimeout),
-      Number, defaults.requestTimeout)
+      Number)
     .option('-u, --url <url>',
-      cliutils.getUsage('URL of the CouchDB/Cloudant server', defaults.url),
-      defaults.url)
+      cliutils.getUsage('URL of the CouchDB/Cloudant server', defaults.url))
     .parse(process.argv);
 
-  // Special case the iamTokenUrl which is only an env var
-  program.iamTokenUrl = defaults.iamTokenUrl;
+  // Remove defaults that don't apply when using shallow mode
+  if (program.mode === 'shallow' || envVarOptions.mode === 'shallow') {
+    delete defaults.parallelism;
+    delete defaults.log;
+    delete defaults.resume;
+  }
 
-  if (program.resume && (program.log === defaults.log)) {
+  // Apply the options in order so that the CLI overrides env vars and env variables
+  // override defaults.
+  const opts = Object.assign({}, defaults, envVarOptions, program);
+
+  if (opts.resume && (opts.log === defaults.log)) {
     // If resuming and the log file arg is the newly generated tmp name from defaults then we know that --log wasn't specified.
     // We have to do this check here for the CLI case because of the default.
     error.terminationCallback(new error.BackupError('NoLogFileName', 'To resume a backup, a log file must be specified'));
   }
 
-  return program;
+  return opts;
 }
 
 function parseRestoreArgs() {
   var program = require('commander');
 
-  var defaults = config.cliDefaults();
-  config.applyEnvironmentVariables(defaults);
+  // Option CLI defaults
+  const defaults = config.cliDefaults();
+
+  // Options set by environment variables
+  const envVarOptions = {};
+  config.applyEnvironmentVariables(envVarOptions);
 
   program
     .version(pkg.version)
@@ -85,28 +97,26 @@ function parseRestoreArgs() {
     .usage('[options...]')
     .option('-b, --buffer-size <n>',
       cliutils.getUsage('number of documents restored at once', defaults.bufferSize),
-      Number, defaults.bufferSize)
+      Number)
     .option('-d, --db <db>',
-      cliutils.getUsage('name of the new, existing database to restore to', defaults.db),
-      defaults.db)
+      cliutils.getUsage('name of the new, existing database to restore to', defaults.db))
     .option('-k, --iam-api-key <API key>',
-      cliutils.getUsage('IAM API key to access the Cloudant server'),
-      defaults.iamApiKey)
+      cliutils.getUsage('IAM API key to access the Cloudant server'))
     .option('-p, --parallelism <n>',
       cliutils.getUsage('number of HTTP requests to perform in parallel when restoring a backup', defaults.parallelism),
-      Number, defaults.parallelism)
+      Number)
     .option('-t, --request-timeout <n>',
       cliutils.getUsage('milliseconds to wait for a response to a HTTP request before retrying the request', defaults.requestTimeout),
-      Number, defaults.requestTimeout)
+      Number)
     .option('-u, --url <url>',
-      cliutils.getUsage('URL of the CouchDB/Cloudant server', defaults.url),
-      defaults.url)
+      cliutils.getUsage('URL of the CouchDB/Cloudant server', defaults.url))
     .parse(process.argv);
 
-  // Special case the iamTokenUrl which is only an env var
-  program.iamTokenUrl = defaults.iamTokenUrl;
+  // Apply the options in order so that the CLI overrides env vars and env variables
+  // override defaults.
+  const opts = Object.assign({}, defaults, envVarOptions, program);
 
-  return program;
+  return opts;
 }
 
 module.exports = {